Feat: fontface 컴포넌트화 및 파싱 유틸 추가
