+++ client/src/components/InjectFontFace/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+interface fontFaceProps {
+  fontFace: string;
+}
+
+export default function InjectFontFace({ fontFace }: fontFaceProps) {
+  const style = React.createElement("style", null, fontFace);
+  return style;
+}
